requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
av==15.0.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
datasets==4.0.0
debugpy==1.8.15
decorator==5.2.1
dill==0.3.8
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.9.0
google-api-core==2.27.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
ics
idna==3.10
ipykernel==6.30.1
ipython==8.37.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==2.12.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.77
langchain-google-genai==2.1.12
langchain-openai==0.3.34
langchain-text-splitters==0.3.11
langgraph==0.6.8
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.31
libclang==18.1.1
lightning==2.5.4
lightning-utilities==0.15.2
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
namex==0.1.0
narwhals==2.0.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
omegaconf==2.3.0
openai==2.1.0
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-lightning==2.5.4
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
stack-data==0.6.3
streamlit==1.47.1
submitit==1.5.3
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.11.0
timm==1.0.19
tokenizers==0.21.4
toml==0.10.2
torch==2.8.0+cpu
torchaudio==2.8.0+cpu
torchmetrics==1.8.2
torchvision==0.23.0+cpu
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.1
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.14.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0

